+++ tools/linters/package.json
@@ -1,7 +1,8 @@
 {
-  "name": "linters",
+  "name": "@cdklabs/linters",
   "version": "1.0.0",
-  "license": "ISC",
+  "license": "Apache-2.0",
+  "private": true,
   "dependencies": {
     "markdownlint-cli": "^0.35.0"
   }
